Crate problem053

Problem 53 (Combinatoric selections)

Solution detail

We can use two facts about binomial coefficients to speed up the search:

nCr is increasing for 0 ≤ r ≤ n / 2

nCn-r = nCr

This means that, for each n, we can iterate up through the values 0 ≤ r ≤ n / 2, and when we find nCr > 1,000,000 for the first time, this means that there are n - 2r + 1 values of r which cause the binomial coefficient to exceed one million.

The values of nCr for increasing r can be calculated using the relation nCr+1 = nCr × (n - r) / (r + 1)

Structs

BinomialCoefficientsWithNumerator
A structure which will allow iteration over binomial coefficients.

Functions

answer
Solve the problem, returning the answer as a String

binomial_coefficients_lower_half
Returns an iterator of the binomials coefficients n choose r, with 0 ≤ r ≤ n / 2.

first_binomial_coefficient_to_exceed
Returns the smallest r for which n choose r exceeds the given threshold, if it exists.

main

solve
Find the number of binomial coefficients n choose r, with n not exceeding the given limit, whose value is greater than the given threshold.